arduino/python-reads-arduino/requirements.txt
# Python requirements.
pyserial
